requirements.txt
kubernetes
robotframework
# for example testcases
robotframework-requests